# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Academy",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "parameters": [
                    {
                        "name": "processTypeId",
                        "in": "path",
                        "description": "process type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Authentication required!"
                    },
                    "403": {
                        "description": "Forbidden request!"
                    },
                    "422": {
                        "description": "Something went wrong!"
                    },
                    "500": {
                        "description": "Internal server error!"
                    }
                }
            }
        },
        "/api/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get me",
                "description": "Get me",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Succes return",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "LoginRequest",
                "description": "LoginRequest",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "email",
                        "type": "string",
                        "example": "test@elek.com"
                    },
                    "password": {
                        "description": "password",
                        "type": "string",
                        "example": "1234"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "id": {
                        "description": "id of user",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "name of user",
                        "type": "string",
                        "example": "Elek"
                    },
                    "role": {
                        "description": "role of user",
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}